requirements.txt
docopts
requests
